refactor api services and models
