+++ frontend/src/api/index.ts
@@ -1,0 +1,6 @@
+export { default as UserService } from './User.service';
+export { default as TeamService } from './Team.service';
+export { default as ProjectService } from './Project.service';
+export { default as SheetService } from './Sheet.service';
+export { default as TeamProjectService } from './TeamProject.service';
+export * from './gradesProcessing';
