+++ frontend/src/models/Grade.model.ts
@@ -1,6 +1,3 @@
-import api from './api';
-
-
 export interface Grades {
     [grade: string]: number;
 };
@@ -26,15 +23,3 @@ export interface GradeSheetData {
         grades: Grades;
     }[];
 };
-
-export default async function getMentorSheets(mentorId?: string): Promise<GradeSheetData[] | null> {
-    let gradeSheetsRes;
-    console.log({mentorId});
-    try {
-        gradeSheetsRes = await api.get<GradeSheetData[]>(`/mentors/${mentorId}/grade/sheets`);
-    } catch(err) {
-        return null;
-    }
-    console.log(gradeSheetsRes);
-    return gradeSheetsRes.data;
-}
