+++ frontend/src/models/Section.model.ts
@@ -1,0 +1,8 @@
+export interface SectionData {
+    _id: string;
+    name: string;
+    startDate: string;
+    endDate: string;
+    referenceProjectId: string;
+    description: string;
+}
