+++ frontend/src/models/Team.model.ts
@@ -1,9 +1,14 @@
+import { User, TeamProject } from './';
+import { UserData } from './User.model';
+
 export interface TeamData {
     _id: string;
     mentor: {
+        _id: string,
         name: string;
         surname: string;
     } | null;
+    users: UserData[],
     course: {
         name: string;
     }
@@ -14,4 +19,16 @@ export interface Team {
     name: string;
     surname: string;
     courseName: string;
+}
+
+export interface TeamInfo {
+    id: string;
+    mentor: {
+      name: string;
+      surname: string;
+    };
+    users: User[];
+    projects: TeamProject[];
+    teamAvgGrade: number;
+    maxPoints: number;
 }
