+++ frontend/src/models/index.ts
@@ -1,0 +1,7 @@
+export * from './Course.model';
+export * from './Grade.model';
+export * from './Project.model';
+export * from './TeamProject.model';
+export * from './Team.model';
+export * from './User.model';
+export * from './Section.model';
